feat: implement Delete
