+++ driver/driver_test.go
@@ -1257,3 +1257,1196 @@ func TestDrop(t *testing.T) {
 		t.Fatalf("cannot drop table: %s", err)
 	}
 }
+
+func TestTrunc(t *testing.T) {
+	db, err := sql.Open("ramsql", "TestTrunc")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT, email TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('id', 'email') VALUES (2, 'bar@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('id', 'email') VALUES (1, 'foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	rows, err := db.Query("SELECT * FROM account WHERE 1")
+	if err != nil {
+		t.Fatalf("sql.Query error : %s", err)
+	}
+
+	columns, err := rows.Columns()
+	if err != nil {
+		t.Fatalf("rows.Column : %s", err)
+		return
+	}
+
+	if len(columns) != 2 {
+		t.Fatalf("Expected 2 columns, got %d", len(columns))
+	}
+
+	res, err := db.Exec("DELETE FROM account")
+	if err != nil {
+		t.Fatalf("Cannot truncate table: %s", err)
+	}
+
+	affectedRows, err := res.RowsAffected()
+	if err != nil {
+		t.Fatalf("Cannot fetch affected rows: %s", err)
+	}
+
+	if affectedRows != 2 {
+		t.Fatalf("Expected 2 rows affected, got %d", affectedRows)
+	}
+
+}
+
+func TestDelete(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestDelete")
+	if err != nil {
+		t.Fatalf("sql.Open: %s", err)
+	}
+	defer db.Close()
+
+	init := []string{
+		`CREATE TABLE queue (id BIGSERIAL, data TEXT)`,
+		`INSERT INTO queue (data) VALUES ($$foo$$)`,
+		`INSERT INTO queue (data) VALUES ($$bar$$)`,
+		`INSERT INTO queue (data) VALUES ($$foobar$$)`,
+		`INSERT INTO queue (data) VALUES ($$barfoo$$)`,
+		`INSERT INTO queue (data) VALUES ($$ok$$)`,
+	}
+	for _, q := range init {
+		_, err := db.Exec(q)
+		if err != nil {
+			t.Fatalf("Cannot initialize test: %s", err)
+		}
+	}
+
+	query := `DELETE FROM queue WHERE id = $1`
+
+	// Delete last
+	_, err = db.Exec(query, 5)
+	if err != nil {
+		t.Fatalf("cannot delete last: %s", err)
+	}
+
+	// Delete first
+	_, err = db.Exec(query, 1)
+	if err != nil {
+		t.Fatalf("cannot delete first: %s", err)
+	}
+
+	// Delete middle
+	_, err = db.Exec(query, 3)
+	if err != nil {
+		t.Fatalf("cannot delete in the midde: %s", err)
+	}
+
+	_, err = db.Exec(query, 4)
+	if err != nil {
+		t.Fatalf("cannot delete row: %s", err)
+	}
+
+	// Delete last
+	_, err = db.Exec(query, 2)
+	if err != nil {
+		t.Fatalf("cannot delete last row in table: %s", err)
+	}
+}
+
+func TestDeleteAnd(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestDeleteAnd")
+	if err != nil {
+		t.Fatalf("sql.Open: %s", err)
+	}
+	defer db.Close()
+
+	init := []string{
+		`CREATE TABLE foo (id BIGSERIAL, bar_id INT, toto_id INT)`,
+		`INSERT INTO foo (bar_id, toto_id) VALUES (2, 3)`,
+		`INSERT INTO foo (bar_id, toto_id) VALUES (4, 32)`,
+		`INSERT INTO foo (bar_id, toto_id) VALUES (5, 33)`,
+		`INSERT INTO foo (bar_id, toto_id) VALUES (6, 4)`,
+	}
+	for _, q := range init {
+		_, err := db.Exec(q)
+		if err != nil {
+			t.Fatalf("Cannot initialize test: %s", err)
+		}
+	}
+
+	query := `DELETE FROM foo WHERE bar_id = $1 AND toto_id = $2`
+	_, err = db.Exec(query, 4, 32)
+	if err != nil {
+		t.Fatalf("cannot delete: %s", err)
+	}
+
+	n := 0
+	query = `SELECT bar_id FROM foo WHERE 1=1`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("cannot query foo: %s", err)
+	}
+	for rows.Next() {
+		n++
+	}
+	if n != 3 {
+		t.Fatalf("Expected 3 values, got %d", n)
+	}
+}
+
+func TestInsertSingle(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertASingle")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE cat (id INT AUTOINCREMENT, breed TEXT, name TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	result, err := db.Exec("INSERT INTO cat (breed, name) VALUES ('indeterminate', 'Uhura')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		t.Fatalf("Cannot check rows affected: %s", err)
+	}
+	if rowsAffected != 1 {
+		t.Fatalf("Expected to affect 1 row, affected %v", rowsAffected)
+	}
+
+	insertedId, err := result.LastInsertId()
+	if err != nil {
+		t.Fatalf("Cannot check last inserted ID: %s", err)
+	}
+
+	row := db.QueryRow("SELECT breed, name FROM cat WHERE id = ?", insertedId)
+	if row == nil {
+		t.Fatalf("sql.Query failed")
+	}
+
+	var breed string
+	var name string
+	err = row.Scan(&breed, &name)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if breed != "indeterminate" || name != "Uhura" {
+		t.Fatalf("Expected breed 'indeterminate' and name 'Uhura', got breed '%v' and name '%v'", breed, name)
+	}
+}
+
+func TestInsertSingleReturning(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertSingleReturning")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE cat (id INT AUTOINCREMENT, breed TEXT, name TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	rows, err := db.Query("INSERT INTO cat (breed, name) VALUES ('indeterminate', 'Nala') RETURNING id")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+	defer rows.Close()
+
+	hasRow := rows.Next()
+	if !hasRow {
+		t.Fatalf("Did not return a row: %s", err)
+	}
+
+	var id int
+	err = rows.Scan(&id)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if id != 1 {
+		t.Fatalf("Expected id 1, got id %v", id)
+	}
+
+	hasRow = rows.Next()
+	if hasRow {
+		t.Fatalf("Returned more than one row: %s", err)
+	}
+}
+
+func TestInsertSingleReturningUint(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertSingleReturningUint")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE cat (id BIGSERIAL PRIMARY KEY, breed TEXT, name TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	rows, err := db.Query("INSERT INTO cat (breed, name) VALUES ('indeterminate', 'Nala') RETURNING id")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+	defer rows.Close()
+
+	hasRow := rows.Next()
+	if !hasRow {
+		t.Fatalf("Did not return a row: %s", err)
+	}
+
+	var id uint
+	err = rows.Scan(&id)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if id != 1 {
+		t.Fatalf("Expected id 1, got id %v", id)
+	}
+
+	hasRow = rows.Next()
+	if hasRow {
+		t.Fatalf("Returned more than one row: %s", err)
+	}
+}
+
+func TestInsertWithMissingValue(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertWithMissingValue")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec(`
+		CREATE TABLE user (
+			id INT AUTOINCREMENT,
+			email TEXT DEFAULT 'example@example.com',
+			name TEXT
+		)
+	`)
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	result, err := db.Exec("INSERT INTO user (name) VALUES ('Bob')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table user: %s", err)
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		t.Fatalf("Cannot check rows affected: %s", err)
+	}
+	if rowsAffected != 1 {
+		t.Fatalf("Expected to affect 1 row, affected %v", rowsAffected)
+	}
+
+	insertedId, err := result.LastInsertId()
+	if err != nil {
+		t.Fatalf("Cannot check last inserted ID: %s", err)
+	}
+
+	row := db.QueryRow("SELECT email, name FROM user WHERE id = ?", insertedId)
+	if row == nil {
+		t.Fatalf("sql.Query failed")
+	}
+
+	var email string
+	var name string
+	err = row.Scan(&email, &name)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if email != "example@example.com" || name != "Bob" {
+		t.Fatalf("Expected email 'example@example.com' and name 'Bob', got email '%v' and name '%v'", email, name)
+	}
+}
+
+func TestInsertMultiple(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertMultiple")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE cat (id INT AUTOINCREMENT, breed TEXT, name TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	result, err := db.Exec("INSERT INTO cat (breed, name) VALUES ('persian', 'Mozart'), ('persian', 'Danton')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		t.Fatalf("Cannot check rows affected: %s", err)
+	}
+	if rowsAffected != 2 {
+		t.Fatalf("Expected to affect 2 rows, affected %v", rowsAffected)
+	}
+}
+
+func TestInsertMultipleReturning(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestInsertMultipleReturning")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE cat (id INT AUTOINCREMENT, breed TEXT, name TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	rows, err := db.Query("INSERT INTO cat (breed, name) VALUES ('indeterminate', 'Spock'), ('indeterminate', 'Belanna') RETURNING id")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+	defer rows.Close()
+
+	hasRow := rows.Next()
+	if !hasRow {
+		t.Fatalf("Did not return a row: %s", err)
+	}
+
+	var id int
+	err = rows.Scan(&id)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+	if id != 1 {
+		t.Fatalf("Expected id 1, got id %v", id)
+	}
+
+	hasRow = rows.Next()
+	if !hasRow {
+		t.Fatalf("Did not return a row: %s", err)
+	}
+
+	err = rows.Scan(&id)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+	if id != 2 {
+		t.Fatalf("Expected id 2, got id %v", id)
+	}
+
+	hasRow = rows.Next()
+	if hasRow {
+		t.Fatalf("Returned more than two rows: %s", err)
+	}
+}
+
+func TestJoinOrderBy(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestJoinOrderBy")
+	if err != nil {
+		t.Fatalf("sql.Open: %s", err)
+	}
+	defer db.Close()
+
+	init := []string{
+		`CREATE TABLE user (id BIGSERIAL, name TEXT)`,
+		`CREATE TABLE address (id BIGSERIAL, user_id INT, value TEXT)`,
+		`INSERT INTO user (name) VALUES ($$riri$$)`,
+		`INSERT INTO user (name) VALUES ($$fifi$$)`,
+		`INSERT INTO user (name) VALUES ($$loulou$$)`,
+		`INSERT INTO address (user_id, value) VALUES (1, 'rue du puit')`,
+		`INSERT INTO address (user_id, value) VALUES (1, 'rue du désert')`,
+		`INSERT INTO address (user_id, value) VALUES (3, 'rue du chemin')`,
+		`INSERT INTO address (user_id, value) VALUES (2, 'boulevard du con')`,
+	}
+	for _, q := range init {
+		_, err := db.Exec(q)
+		if err != nil {
+			t.Fatalf("Cannot initialize test: %s", err)
+		}
+	}
+
+	query := `SELECT user.name, address.value
+			FROM user
+			JOIN address ON address.user_id = user.id
+			WHERE user.id = $1
+			ORDER BY address.value ASC`
+	rows, err := db.Query(query, 1)
+	if err != nil {
+		t.Fatalf("Cannot select with joined order by: %s", err)
+	}
+	defer rows.Close()
+
+	n := 0
+	for rows.Next() {
+		n++
+	}
+	if n != 2 {
+		t.Fatalf("Expected 2 rows, got %d", n)
+	}
+}
+
+func TestMultipleJoin(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestMultipleJoin")
+	if err != nil {
+		t.Fatalf("sql.Open: %s", err)
+	}
+	defer db.Close()
+
+	init := []string{
+		`CREATE TABLE user (id BIGSERIAL, name TEXT)`,
+		`CREATE TABLE address (id BIGSERIAL, user_id INT, value TEXT)`,
+		`CREATE TABLE user_group (id BIGSERIAL, user_id INT, name TEXT)`,
+		`INSERT INTO user (name) VALUES ($$riri$$)`,
+		`INSERT INTO user (name) VALUES ($$fifi$$)`,
+		`INSERT INTO user (name) VALUES ($$loulou$$)`,
+		`INSERT INTO address (user_id, value) VALUES (1, 'rue du puit')`,
+		`INSERT INTO address (user_id, value) VALUES (1, 'rue du désert')`,
+		`INSERT INTO address (user_id, value) VALUES (3, 'rue du chemin')`,
+		`INSERT INTO address (user_id, value) VALUES (2, 'boulevard du con')`,
+		`INSERT INTO user_group (user_id, name) VALUES (1, 'toto')`,
+		`INSERT INTO user_group (user_id, name) VALUES (2, 'toto')`,
+		`INSERT INTO user_group (user_id, name) VALUES (3, 'lonely')`,
+		`INSERT INTO user_group (user_id, name) VALUES (1, 'cowboy')`,
+	}
+	for _, q := range init {
+		_, err := db.Exec(q)
+		if err != nil {
+			t.Fatalf("Cannot initialize test: %s", err)
+		}
+	}
+
+	query := `SELECT user.name, address.value
+			FROM user
+			JOIN address ON address.user_id = user.id
+			JOIN user_group ON user_group.user_id = address.user_id
+			WHERE user_group.name = $1
+			ORDER BY address.value ASC`
+	rows, err := db.Query(query, "toto")
+	if err != nil {
+		t.Fatalf("Cannot select with 3 tables joined: %s", err)
+	}
+	defer rows.Close()
+
+	n := 0
+	for rows.Next() {
+		n++
+	}
+	if n != 3 {
+		t.Fatalf("Expected 3 rows, got %d", n)
+	}
+
+}
+
+func TestJoinGroup(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestJoinGroup")
+	if err != nil {
+		t.Fatalf("sql.Open: %s", err)
+	}
+	defer db.Close()
+
+	init := []string{
+		`CREATE TABLE user (id BIGSERIAL, name TEXT)`,
+		`CREATE TABLE group (id BIGSERIAL, name TEXT)`,
+		`CREATE TABLE user_group (id BIGSERIAL, user_id INT, group_id INT)`,
+		`INSERT INTO user (name) VALUES ($$riri$$)`,
+		`INSERT INTO user (name) VALUES ($$fifi$$)`,
+		`INSERT INTO user (name) VALUES ($$loulou$$)`,
+		`INSERT INTO group (name) VALUES ('cowboys')`,
+		`INSERT INTO group (name) VALUES ('troopers')`,
+		`INSERT INTO group (name) VALUES ('toys')`,
+		`INSERT INTO user_group (user_id, group_id) VALUES (1, 1)`,
+		`INSERT INTO user_group (user_id, group_id) VALUES (2, 1)`,
+	}
+	for _, q := range init {
+		_, err := db.Exec(q)
+		if err != nil {
+			t.Fatalf("Cannot initialize test: %s", err)
+		}
+	}
+
+	query := `SELECT user.name FROM user
+			JOIN user_group ON user.id = user_group.user_id
+			WHERE user_group.group_id = $1
+			ORDER BY user.name ASC`
+	rows, err := db.Query(query, 1)
+	if err != nil {
+		t.Fatalf("Cannot select joined: %s", err)
+	}
+	defer rows.Close()
+
+	n := 0
+	for rows.Next() {
+		n++
+	}
+	if n != 2 {
+		t.Fatalf("Expected 2 rows, got %d", n)
+	}
+
+}
+
+func TestOrderByInt(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestOrderByInt")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE user (name TEXT, surname TEXT, age INT);`,
+		`INSERT INTO user (name, surname, age) VALUES (Foo, Bar, 20);`,
+		`INSERT INTO user (name, surname, age) VALUES (John, Doe, 32);`,
+		`INSERT INTO user (name, surname, age) VALUES (Jane, Doe, 33);`,
+		`INSERT INTO user (name, surname, age) VALUES (Joe, Doe, 10);`,
+		`INSERT INTO user (name, surname, age) VALUES (Homer, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Marge, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Bruce, Wayne, 3333);`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s", err)
+		}
+	}
+
+	query := `SELECT age FROM user WHERE surname = Wayne OR surname = Doe ORDER BY age DESC`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("Cannot select and order by age: %s", err)
+	}
+	defer rows.Close()
+
+	var age, last, size int64
+	last = 4000
+	// Now every time 'age' should be less than 'last'
+	for rows.Next() {
+		err = rows.Scan(&age)
+		if err != nil {
+			t.Fatalf("Cannot scan age: %s", err)
+		}
+		if age > last {
+			t.Fatalf("Got %d previously and now %d", last, age)
+		}
+		last = age
+		size++
+	}
+
+	if size != 4 {
+		t.Fatalf("Expecting 4 rows here, got %d", size)
+	}
+
+	query = `SELECT age FROM user ORDER BY age ASC`
+	rows, err = db.Query(query)
+	if err != nil {
+		t.Fatalf("cannot order by age: %s\n", err)
+	}
+
+	size = 0
+	last = 0
+	// Now 'last' should be less than current 'age'
+	for rows.Next() {
+		err = rows.Scan(&age)
+		if err != nil {
+			t.Fatalf("Cannot scan age: %s", err)
+		}
+		if last > age {
+			t.Fatalf("Got %d previously and now %d", last, age)
+		}
+		last = age
+		size++
+	}
+
+	if size != 7 {
+		t.Fatalf("Expecting 7 rows, got %d", size)
+	}
+}
+
+func TestOrderByString(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestOrderByString")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE user (name TEXT, surname TEXT, age INT);`,
+		`INSERT INTO user (name, surname, age) VALUES (Foo, Bar, 20);`,
+		`INSERT INTO user (name, surname, age) VALUES (John, Doe, 32);`,
+		`INSERT INTO user (name, surname, age) VALUES (Jane, Doe, 33);`,
+		`INSERT INTO user (name, surname, age) VALUES (Joe, Doe, 10);`,
+		`INSERT INTO user (name, surname, age) VALUES (Homer, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Marge, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Bruce, Wayne, 3333);`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s", err)
+		}
+	}
+
+	query := `SELECT name, surname FROM user WHERE surname = $1 ORDER BY name ASC`
+	rows, err := db.Query(query, "Doe")
+	if err != nil {
+		t.Fatalf("Cannot select and order by age: %s", err)
+	}
+
+	var names []string
+	var name, surname string
+	for rows.Next() {
+		err = rows.Scan(&name, &surname)
+		if err != nil {
+			t.Fatalf("Cannot scan row: %s\n", err)
+		}
+		if surname != "Doe" {
+			t.Fatalf("Didn't expect surname being %s", surname)
+		}
+		names = append(names, name)
+	}
+
+	if len(names) != 3 {
+		t.Fatalf("Expected 3 rows, not %d", len(names))
+	}
+
+	if names[0] != "Jane" {
+		t.Fatalf("Wanted Jane, got %s", names[0])
+	}
+
+	if names[1] != "Joe" {
+		t.Fatalf("Wanted Joe, got %s", names[1])
+	}
+
+	if names[2] != "John" {
+		t.Fatalf("Wanted John, got %s", names[2])
+	}
+
+}
+
+func TestOrderByLimit(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestOrderByLimit")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE user (name TEXT, surname TEXT, age INT);`,
+		`INSERT INTO user (name, surname, age) VALUES (Foo, Bar, 20);`,
+		`INSERT INTO user (name, surname, age) VALUES (John, Doe, 32);`,
+		`INSERT INTO user (name, surname, age) VALUES (Jane, Doe, 33);`,
+		`INSERT INTO user (name, surname, age) VALUES (Joe, Doe, 10);`,
+		`INSERT INTO user (name, surname, age) VALUES (Homer, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Marge, Simpson, 40);`,
+		`INSERT INTO user (name, surname, age) VALUES (Bruce, Wayne, 3333);`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s", err)
+		}
+	}
+
+	log.SetLevel(log.DebugLevel)
+	defer log.SetLevel(log.InfoLevel)
+	query := `SELECT age FROM user WHERE 1=1 ORDER BY age DESC LIMIT 2`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("Cannot limit select: %s", err)
+	}
+
+	var age, last, size int64
+	last = 4000
+	// Now every time 'age' should be less than 'last'
+	for rows.Next() {
+		err = rows.Scan(&age)
+		if err != nil {
+			t.Fatalf("Cannot scan age: %s", err)
+		}
+		if age > last {
+			t.Fatalf("Got %d previously and now %d", last, age)
+		}
+		last = age
+		size++
+	}
+
+	if size != 2 {
+		t.Fatalf("Expecting 2 rows here, got %d", size)
+	}
+}
+
+func TestOrderByIntEmpty(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestOrderByIntEmpty")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE user (name TEXT, surname TEXT, age INT);`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s", err)
+		}
+	}
+
+	query := `SELECT age FROM user WHERE surname = Wayne OR surname = Doe ORDER BY age DESC`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("Cannot select and order by age: %s", err)
+	}
+	defer rows.Close()
+
+}
+
+func TestOrderByMultipleStrings(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestOrderByMultipleStrings")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+
+	defer func() {
+		if err := db.Close(); err != nil {
+			t.Error(err)
+		}
+	}()
+
+	batch := []string{
+		`CREATE TABLE user (name TEXT, surname TEXT, age INT64, savings DECIMAL);`,
+		`INSERT INTO user (name, surname, age, savings) VALUES (Joe, Angel, 11, 125.215);`,
+		`INSERT INTO user (name, surname, age, savings) VALUES (Joe, Angel, 11, 1.1);`,
+		`INSERT INTO user (name, surname, age, savings) VALUES (Joe, Zebra, 32, 0.921);`,
+		`INSERT INTO user (name, surname, age, savings) VALUES (Anna, Angel, 33, 0);`,
+		`INSERT INTO user (name, surname, age, savings) VALUES (Anna, Zebra, 9, 25);`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s", err)
+		}
+	}
+
+	query := `SELECT name, surname, age, savings FROM user ORDER BY name ASC, age DESC, savings ASC`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("Cannot select and order by age: %s", err)
+	}
+
+	exp := [][]string{
+		{"Anna", "33", "0"},
+		{"Anna", "9", "25"},
+		{"Joe", "32", "0.921"},
+		{"Joe", "11", "1.1"},
+		{"Joe", "11", "125.215"},
+	}
+	var (
+		got           [][]string
+		name, surname string
+		age           int64
+		savings       float64
+	)
+	for rows.Next() {
+		err = rows.Scan(&name, &surname, &age, &savings)
+		if err != nil {
+			t.Fatalf("Cannot scan row: %s\n", err)
+		}
+		got = append(got, []string{name, fmt.Sprint(age), fmt.Sprint(savings)})
+	}
+
+	if len(exp) != len(got) {
+		t.Fatalf("length mismatch, expected %d but got %d", len(exp), len(got))
+	}
+	for i := 0; i < len(exp); i++ {
+		for j := 0; j < len(exp[i]); j++ {
+			if exp[i][j] != got[i][j] {
+				t.Fatalf("data mismatch at %d/%d, expected %v but got %v", i, j, exp[i][j], got[i][j])
+			}
+		}
+	}
+}
+
+func TestSelectNoOp(t *testing.T) {
+	db, err := sql.Open("ramsql", "TestSelectNoOp")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE account (id BIGSERIAL, email TEXT)`,
+		`INSERT INTO account (email) VALUES ("foo@bar.com")`,
+		`INSERT INTO account (email) VALUES ("bar@bar.com")`,
+		`INSERT INTO account (email) VALUES ("foobar@bar.com")`,
+		`INSERT INTO account (email) VALUES ("babar@bar.com")`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s\n", err)
+		}
+	}
+
+	query := `SELECT * from account WHERE 1 = 1`
+	rows, err := db.Query(query)
+	if err != nil {
+		t.Fatalf("cannot create table: %s", err)
+	}
+
+	nb := 0
+	for rows.Next() {
+		nb++
+	}
+
+	if nb != 4 {
+		t.Fatalf("Expected 4 rows, got %d", nb)
+	}
+
+}
+
+func TestSelect(t *testing.T) {
+	db, err := sql.Open("ramsql", "TestSelect")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT, email TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('id', 'email') VALUES (2, 'bar@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('id', 'email') VALUES (1, 'foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	rows, err := db.Query("SELECT * FROM account WHERE email = $1", "foo@bar.com")
+	if err != nil {
+		t.Fatalf("sql.Query error : %s", err)
+	}
+
+	columns, err := rows.Columns()
+	if err != nil {
+		t.Fatalf("rows.Column : %s", err)
+		return
+	}
+
+	if len(columns) != 2 {
+		t.Fatalf("Expected 2 columns, got %d", len(columns))
+	}
+
+	row := db.QueryRow("SELECT * FROM account WHERE email = $1", "foo@bar.com")
+	if row == nil {
+		t.Fatalf("sql.QueryRow error")
+	}
+
+	var email string
+	var id int
+	err = row.Scan(&id, &email)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if id != 1 {
+		t.Fatalf("Expected id = 1, got %d", id)
+	}
+
+	if email != "foo@bar.com" {
+		t.Fatalf("Expected email = <foo@bar.com>, got <%s>", email)
+	}
+
+	err = db.Close()
+	if err != nil {
+		t.Fatalf("sql.Close : Error : %s\n", err)
+	}
+
+}
+
+func TestCount(t *testing.T) {
+	db, err := sql.Open("ramsql", "TestCount")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	batch := []string{
+		`CREATE TABLE account (id BIGSERIAL, email TEXT)`,
+		`INSERT INTO account (email) VALUES ("foo@bar.com")`,
+		`INSERT INTO account (email) VALUES ("bar@bar.com")`,
+		`INSERT INTO account (email) VALUES ("foobar@bar.com")`,
+		`INSERT INTO account (email) VALUES ("babar@bar.com")`,
+	}
+
+	for _, b := range batch {
+		_, err = db.Exec(b)
+		if err != nil {
+			t.Fatalf("sql.Exec: Error: %s\n", err)
+		}
+	}
+
+	var count int64
+	err = db.QueryRow(`SELECT COUNT(*) FROM account WHERE 1=1`).Scan(&count)
+	if err != nil {
+		t.Fatalf("cannot select COUNT of account: %s\n", err)
+	}
+
+	if count != 4 {
+		t.Fatalf("Expected count to be 4, not %d", count)
+	}
+
+	err = db.QueryRow(`SELECT COUNT(i_dont_exist_lol) FROM account WHERE 1=1`).Scan(&count)
+	if err == nil {
+		t.Fatalf("Expected an error from a non existing attribute")
+	}
+
+}
+
+func TestUpdateSimple(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestUpdateSimple")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT AUTOINCREMENT, email TEXT)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('leon@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("UPDATE account SET email = 'roger@gmail.com' WHERE id = 2")
+	if err != nil {
+		t.Fatalf("Cannot update table account: %s", err)
+	}
+
+	row := db.QueryRow("SELECT * FROM account WHERE id = 2")
+	if row == nil {
+		t.Fatalf("sql.Query failed")
+	}
+
+	var email string
+	var id int
+	err = row.Scan(&id, &email)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+
+	if email != "roger@gmail.com" {
+		t.Fatalf("Expected email 'roger@gmail.com', got '%s'", email)
+	}
+}
+
+func TestUpdateIsNull(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestUpdateIsNull")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT AUTOINCREMENT, email TEXT, creation_date TIMESTAMP WITH TIME ZONE)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('leon@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	res, err := db.Exec("UPDATE account SET email = 'roger@gmail.com', creation_date = $1 WHERE id = 2 AND creation_date IS NULL", time.Now())
+	if err != nil {
+		t.Fatalf("Cannot update table account: %s", err)
+	}
+
+	ra, err := res.RowsAffected()
+	if err != nil {
+		t.Fatalf("Cannot check number of rows affected: %s", err)
+	}
+	if ra != 1 {
+		t.Fatalf("Expected 1 row, affected. Got %d", ra)
+	}
+
+	rows, err := db.Query(`SELECT id FROM account WHERE creation_date IS NULL`)
+	if err != nil {
+		t.Fatalf("cannot select null columns: %s", err)
+	}
+
+	var n, id int64
+	for rows.Next() {
+		n++
+		err = rows.Scan(&id)
+		if err != nil {
+			t.Fatalf("cannot scan null columns: %s", err)
+		}
+	}
+	rows.Close()
+	if n != 1 {
+		t.Fatalf("Expected 1 rows, got %d", n)
+	}
+
+	rows, err = db.Query(`SELECT id FROM account WHERE creation_date IS NOT NULL`)
+	if err != nil {
+		t.Fatalf("cannot select not null columns: %s", err)
+	}
+
+	n = 0
+	for rows.Next() {
+		n++
+		err = rows.Scan(&id)
+		if err != nil {
+			t.Fatalf("cannot scan null columns: %s", err)
+		}
+	}
+	rows.Close()
+	if n != 1 {
+		t.Fatalf("Expected 1 rows, got %d", n)
+	}
+
+}
+
+func TestUpdateNotNull(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestUpdateNotNull")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT AUTOINCREMENT, email TEXT, creation_date TIMESTAMP WITH TIME ZONE)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('leon@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	_, err = db.Exec("UPDATE account SET email = 'roger@gmail.com' WHERE id = 2 AND creation_date IS NOT NULL")
+	if err != nil {
+		t.Fatalf("Cannot update table account: %s", err)
+	}
+
+}
+
+func TestUpdateToNull(t *testing.T) {
+
+	db, err := sql.Open("ramsql", "TestUpdateToNull")
+	if err != nil {
+		t.Fatalf("sql.Open : Error : %s\n", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec("CREATE TABLE account (id INT AUTOINCREMENT, email TEXT, creation_date TIMESTAMP WITH TIME ZONE)")
+	if err != nil {
+		t.Fatalf("sql.Exec: Error: %s\n", err)
+	}
+
+	_, err = db.Exec("INSERT INTO account ('email') VALUES ('foo@bar.com')")
+	if err != nil {
+		t.Fatalf("Cannot insert into table account: %s", err)
+	}
+
+	row1 := db.QueryRow("SELECT email FROM account WHERE id = 1")
+	if row1 == nil {
+		t.Fatalf("sql.Query failed")
+	}
+
+	var email1 *string
+	err = row1.Scan(&email1)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+	if email1 == nil {
+		t.Fatalf("expected 'foo@bar.com' email, but got NULL")
+	}
+
+	_, err = db.Exec("UPDATE account SET email = NULL WHERE id = 1")
+	if err != nil {
+		t.Fatalf("Cannot update table account: %s", err)
+	}
+
+	row2 := db.QueryRow("SELECT email FROM account WHERE id = 1")
+	if row2 == nil {
+		t.Fatalf("sql.Query failed")
+	}
+
+	var email2 *string
+	err = row2.Scan(&email2)
+	if err != nil {
+		t.Fatalf("row.Scan: %s", err)
+	}
+	if email2 != nil {
+		t.Fatalf("expected NULL email, but got '%v'", *email2)
+	}
+
+}
